refactor: move Ethereum-specific RPC logic from `CompletedTx` to Ethereum indexer
