+++ chain-signatures/node/src/respond_bidirectional.rs
@@ -1,7 +1,5 @@
-use crate::indexer_eth::EthereumClient;
 use crate::protocol::{Chain, IndexedSignRequest};
 use crate::sign_bidirectional::{BidirectionalTxExt, TransactionOutput};
-use alloy::consensus::Transaction;
 use alloy::primitives::Bytes;
 use k256::Scalar;
 use mpc_crypto::ScalarExt;
@@ -9,7 +7,6 @@ use mpc_primitives::{
     BidirectionalTx, RespondBidirectionalSerializedOutput, RespondBidirectionalTx, SerDeserFormat,
     SignArgs, SignId,
 };
-use std::sync::Arc;
 
 const MAGIC_ERROR_PREFIX: [u8; 4] = [0xde, 0xad, 0xbe, 0xef];
 const SOLANA_RESPOND_BIDIRECTIONAL_PATH: &str = "solana response key";
@@ -120,62 +117,13 @@ impl CompletedTx {
             },
         ))
     }
-
-    pub async fn extract_success_tx_output(
-        &self,
-        client: &Arc<EthereumClient>,
-    ) -> anyhow::Result<RespondBidirectionalSerializedOutput> {
-        let tx = &self.tx;
-        let tx_id = self.tx.id.0;
-        let Some(tx_info) = client
-            .as_ref()
-            .get_transaction_by_hash(tx_id.into())
-            .await?
-        else {
-            anyhow::bail!("Failed to fetch transaction {tx_id:?}");
-        };
-
-        // A deployment transaction has no `to`. For a CREATE the trace's
-        // `output` is the deployed runtime bytecode, not ABI return data, so it
-        // cannot be decoded against the output schema — reject it up front.
-        // TODO(#808): support contract deployments.
-        if tx_info.inner.to().is_none() {
-            anyhow::bail!("unsupported contract deployment (CREATE): {tx_id:?}");
-        }
-
-        let data = tx_info.inner.input().clone();
-        let is_contract_call = is_contract_call(&data);
-
-        let trace_output = if is_contract_call {
-            tracing::info!(
-                ?tx_id,
-                "Extracting transaction output via debug_traceTransaction"
-            );
-            Some(client.trace_transaction_output(tx_id.into()).await?)
-        } else {
-            None
-        };
-
-        build_serialized_output(
-            is_contract_call,
-            &tx.output_deserialization_schema,
-            trace_output.as_ref(),
-            tx.source_chain.respond_serialization_format(),
-            &tx.respond_serialization_schema,
-        )
-    }
-}
-
-/// Whether a transaction's calldata represents a contract call.
-fn is_contract_call(input: &Bytes) -> bool {
-    input.len() > 2 && input != &Bytes::from("0x")
 }
 
 /// Decode a transaction's output and re-serialize it for the respond chain.
 ///
 /// `trace_output` is the `debug_traceTransaction` return data, required when
 /// `is_contract_call` is true.
-fn build_serialized_output(
+pub fn build_serialized_output(
     is_contract_call: bool,
     output_deserialization_schema: &[u8],
     trace_output: Option<&Bytes>,
@@ -247,14 +195,14 @@ mod tests {
         Bytes::from(buf.to_vec())
     }
 
-    #[test]
-    fn is_contract_call_detects_calldata() {
-        assert!(!is_contract_call(&Bytes::new()));
-        assert!(!is_contract_call(&Bytes::from(vec![0u8; 2])));
-        assert!(is_contract_call(&Bytes::from(vec![
-            0xa9, 0x05, 0x9c, 0xbb, 0x00
-        ])));
-    }
+    // #[test]
+    // fn is_contract_call_detects_calldata() {
+    //     assert!(!is_contract_call(&Bytes::new()));
+    //     assert!(!is_contract_call(&Bytes::from(vec![0u8; 2])));
+    //     assert!(is_contract_call(&Bytes::from(vec![
+    //         0xa9, 0x05, 0x9c, 0xbb, 0x00
+    //     ])));
+    // }
 
     #[test]
     fn build_serialized_output_decodes_contract_call() {
